Improve the page size retrieval code via sysconf(3)
Machine code that retrieves the page size through
`sysconf(_SC_PAGESIZE)` now behaves consistently across platforms.
The return values of 0 and negative are now treated as invalid.
The `pageSize` members for these Machine objects are now all `size_t`
type (the return value type of `sysconf()` is `long`).

+++ netbsd/NetBSDMachine.h
@@ -12,7 +12,7 @@ in the source distribution for its full text.
 
 #include <kvm.h>
 #include <stdbool.h>
-#include <sys/types.h>
+#include <stddef.h>
 
 #include "Machine.h"
 #include "ProcessTable.h"
@@ -45,8 +45,8 @@ typedef struct NetBSDMachine_ {
    kvm_t* kd;
 
    long fscale;
-   int pageSize;
-   int pageSizeKB;
+   size_t pageSize;
+   size_t pageSizeKB;
 
    CPUData* cpuData;
 } NetBSDMachine;
